# IGNITE-21558 Sql. Remove ExecutionContext dependency from ExpressionFactory