--- a/modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/DynamicPartitionProvider.java
+++ b/modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/DynamicPartitionProvider.java
@@ -55,6 +55,6 @@ public class DynamicPartitionProvider<RowT> implements PartitionProvider<RowT> {
     public List<PartitionWithConsistencyToken> getPartitions(ExecutionContext<RowT> ctx) {
         ExpressionFactory<RowT> expressionFactory = ctx.expressionFactory();
 
-        return PartitionPruningPredicate.prunePartitions(columns, table, expressionFactory, assignments, nodeName);
+        return PartitionPruningPredicate.prunePartitions(ctx, columns, table, expressionFactory, assignments, nodeName);
     }
 }