… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/LogicalRelImplementor.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/ExpressionFactoryImpl.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/FieldGetter.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/JoinPredicateImplementor.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/NoOpFieldGetter.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/PredicateImplementor.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/ProjectionImplementor.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/RexImpTable.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/RexToLixTranslator.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/RowProviderImplementor.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/ScalarImplementor.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/SearchBoundsImplementor.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/SqlComparator.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/SqlJoinPredicate.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/SqlPredicate.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/SqlProjection.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/SqlRowProvider.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/SqlScalar.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/ValuesImplementor.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/agg/AccumulatorsFactory.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/func/TableFunctionRegistryImpl.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/rel/HashAggregateNode.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/rel/SortAggregateNode.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/prepare/KeyValueGetPlan.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/prepare/KeyValueModifyPlan.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/prepare/SelectCountPlan.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/prepare/pruning/PartitionPruningPredicate.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/util/Commons.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/util/IgniteMethod.java
… modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/util/RexUtils.java